+++ generators/chipyard/src/main/scala/config/NoCConfigs.scala
@@ -3,158 +3,186 @@ package chipyard
 import freechips.rocketchip.config.{Config}
 import freechips.rocketchip.diplomacy.{AsynchronousCrossing}
 import freechips.rocketchip.subsystem.{SBUS, MBUS}
-import constellation.channel.{UserVirtualChannelParams}
+import constellation.channel._
 import constellation.routing._
 import constellation.topology._
+import constellation.noc._
+import constellation.soc.{GlobalNoCParams}
 
+import scala.collection.immutable.ListMap
 
 /*
- * This config demonstrates a 4x4 Mesh NoC topology implemented
- * with Constellation. The SystemBus and MemoryBus are mapped to a
- * shared network, while the ControlBus is mapped to an independent
- * narrow network.
- * Mapping of nodes onto the topology is performed via string-matching
- * on the diplomatic names for the edges into each bus. Unfortunately
- * these strings are not generated with obvious names, so the mapping
- * is a little obtuse.
+ * This config demonstrates a SoC architecture with three different
+ * independent interconnects. The SBus interconnect is the main
+ * backbone, handling core traffic to the L2s and peripherals.
+ * The MBus handles requests to the DRAM and external memory.
+ * The CBus handles requests to control devices.
  *
- * As with all other CDE Configs, the configuration is built up from
- * bottom-up.
- *
- * Topology:
+ * Topologies:
  *
  * 12 - 13 - 14 - 15
  *  |    |    |    |
- *  8 -- 9 - 10 - 11
- *  |    |    |    |
- *  4 -- 5 -- 6 -- 7
+ *  8 -- 9 - 10 - 11    0 -- 1 -- 2 -- 3
+ *  |    |    |    |    |              |
+ *  4 -- 5 -- 6 -- 7    7 -- 6 -- 5 -- 4
  *  |    |    |    |
  *  0 -- 1 -- 2 -- 3
  *
  * This table describes the mappings of each edge onto the network
- * Note that L2s are managers (outwards edges) from the sbus, but are
- * clients (inwards edges) into the mbus
- * 
+ *
  * SI/SO: Inward/outward names into sbus
- * MI/MO: Inward/otward names into mbus
+ * MI/MO: Inward/outward names into mbus
  *
  *   |(12)__________|(13)__________|(14)__________|(15)__________|
- *   | DRAM 0       | DRAM 1       | DRAM 2       | DRAM 3       |
- *   | MO:system[0] | MO:system[1] | MO:system[2] | MO:system[3] |
- *   | MO:serdesser |              |              |              |
- *   |              |              |              |              |
+ *   |              | Core 6       | Core 7       |              |
+ *   |              | SI:Core 6    | SI:Core 7    |              |
  *   |(8)___________|(9)___________|(10)__________|(11)__________|
- *   | Core1        | Core2        | Core3        | Core4        |
- *   | SI:Core 1    | SI:Core 2    | SI:Core 3    | SI:Core 4    |
- *   |              |              |              |              |
- *   |              |              |              |              |
+ *   | Core 4       | L2 2         | L2 3         | Core 5       |
+ *   | SI:Core 4    | S0:system[2] | SO:system[3] | SI:Core 5    |
  *   |(4)___________|(5)___________|(6)___________|(7)___________|
- *   |              | L2_0         | L2_1         | Core0+Pbus   |
- *   |              | SO:system[0] | SO:system[1] | SI:Core 0    |
- *   |              | MI:L2[0]     | MI:L2[1]     | SO:pbus      |
- *   |              |              |              |              |
+ *   | Core 2       | L2 0         | L2 1         | Core 3       |
+ *   | SI:Core 2    | SO:system[0] | SO:system[1] | SI:Core 3    |
  *   |(0)___________|(1)___________|(2)___________|(3)___________|
- *   | FBus         | L2_4         | L2_3         |              |
- *   | SI:serial_tl | SO:System[2] | SO:system[3] |              |
- *   |              | MI:L2[2]     | MI:L2[3]     |              |
+ *   | FBus         | Core 0       | Core 1       | Pbus         |
+ *   | SI:serial-tl | SI:Core 0    | SI:Core 1    | SO:pbus      |
+ *   |______________|______________|______________|______________|
+ *
+ *   |(0)___________|(1)___________|(2)___________|(3)___________|
+ *   | DRAM 0       | L2 0         | L2 1         | DRAM 1       |
+ *   | M0:system[0] | MI:L2[0]     | MI:L2[1]     | M0:system[1] |
+ *   | M0:serdesser |              |              |              |
+ *   |______________|______________|______________|______________|
+ *    ||||||||||||||                               ||||||||||||||
+ *   |(7)___________|(6)___________|(5)___________|(4)___________|
+ *   | DRAM 2       | L2 2         | L2 3         | DRAM 3       |
+ *   | M0:system[2] | MI:L2[2]     | MI:L2[3]     | M0:system[3] |
  *   |              |              |              |              |
  *   |______________|______________|______________|______________|
  */
-class BigNoCConfig extends Config(
-  // Map the inwards and outwards edges of the cbus to cbus NoC nodes
-  new constellation.rc.WithCBusNoCGlobalNoCCtrlMapping((i) => i % 16) ++
-  new constellation.rc.WithCbusNoCOutNodeMapping("chipyardPRCI[1]", 7) ++
-  new constellation.rc.WithCbusNoCOutNodeMapping("chipyardPRCI[0]", 7) ++
-  new constellation.rc.WithCbusNoCOutNodeMapping("bootrom", 7) ++
-  new constellation.rc.WithCbusNoCOutNodeMapping("dmInner", 7) ++
-  new constellation.rc.WithCbusNoCOutNodeMapping("clint", 7) ++
-  new constellation.rc.WithCbusNoCOutNodeMapping("plic", 7) ++
-  new constellation.rc.WithCbusNoCOutNodeMapping("pbus", 7) ++
-  new constellation.rc.WithCbusNoCOutNodeMapping("l2", 7) ++ // TODO fix this should be per L2 bank
-  new constellation.rc.WithCbusNoCOutNodeMapping("error", 7) ++
-  new constellation.rc.WithCbusNoCInNodeMapping("", 7) ++
-
-  // Replce the Xbar-based cbus with a NoC based cbus (and force a narrow width NoC)
-  new constellation.rc.WithCbusNoC(explicitWidth = Some(16)) ++
-
-  // Configure the physical resources of the lightweight control interconnect
-  // 4x4 2D mesh with support for 5 nonblocking virtual subnetworks
-  // Cbus -> 5 subnets for TL ABCDE protocol channels (BCE are optimized away)
-  new constellation.routing.WithNNonblockingVirtualNetworks(5) ++
-  new constellation.channel.WithUniformNVirtualChannels(5, UserVirtualChannelParams(4)) ++
-  new constellation.routing.WithTerminalPlaneRouting ++
-  new constellation.routing.WithRoutingRelation(new Mesh2DDimensionOrderedRouting(4, 4)) ++
-  new constellation.topology.WithTerminalPlane ++
-  new constellation.topology.WithTopology(new Mesh2D(4, 4)) ++
-
-  // Indicate that the mbus and sbus nocs should be implemented
-  // by the shared global noc. Global noc payload width needs to be
-  // wide enough to to support 256-b wide sbus width
-  new constellation.rc.WithGlobalNoCWidth(300) ++
-  new constellation.rc.WithMbusGlobalNoC ++
-  new constellation.rc.WithSbusGlobalNoC ++
-  new constellation.noc.WithNoParamValidation ++
-
-  // Configure the physical resources of the main global data interconnect
-  // 4x4 2D mesh with escape-channel routing + support for 10 nonblocking virtual subnetworks
-  // Sbus -> 5 subnets for TL ABCDE protocol channels
-  // Mbus -> 5 subnets for TL ABCDE protocol channels (BCE are optimized away)
-  new constellation.channel.WithUniformNVirtualChannels(13, UserVirtualChannelParams(7)) ++
-  new constellation.routing.WithNNonblockingVirtualNetworksWithSharing(10, 3) ++
-  new constellation.routing.WithTerminalPlaneRouting ++
-  new constellation.routing.WithRoutingRelation(new Mesh2DEscapeRouting(4, 4)) ++
-  new constellation.topology.WithTerminalPlane ++
-  new constellation.topology.WithTopology(new Mesh2D(4, 4)) ++
-
-
-  // Map the inwards and outwards edges of the mbus to topology nodes
-  new constellation.rc.WithMbusNoCOutNodeMapping("serdesser", 12) ++
-  new constellation.rc.WithMbusNoCOutNodeMapping("system[0]", 12) ++
-  new constellation.rc.WithMbusNoCOutNodeMapping("system[1]", 13) ++
-  new constellation.rc.WithMbusNoCOutNodeMapping("system[2]", 14) ++
-  new constellation.rc.WithMbusNoCOutNodeMapping("system[3]", 15) ++
-  new constellation.rc.WithMbusNoCInNodeMapping("L2 InclusiveCache[3]", 2) ++
-  new constellation.rc.WithMbusNoCInNodeMapping("L2 InclusiveCache[2]", 1) ++
-  new constellation.rc.WithMbusNoCInNodeMapping("L2 InclusiveCache[1]", 6) ++
-  new constellation.rc.WithMbusNoCInNodeMapping("L2 InclusiveCache[0]", 5) ++
-
-  // Replace the Xbar-based mbus with a NoC based mbus
-  new constellation.rc.WithMbusNoC ++
-
-  // Map the inwards and outwards edges of the sbus to topology nodes
-  new constellation.rc.WithSbusNoCOutNodeMapping("system[3]", 2) ++
-  new constellation.rc.WithSbusNoCOutNodeMapping("system[2]", 1) ++
-  new constellation.rc.WithSbusNoCOutNodeMapping("system[1]", 6) ++
-  new constellation.rc.WithSbusNoCOutNodeMapping("system[0]", 5) ++
-  new constellation.rc.WithSbusNoCOutNodeMapping("pbus"     , 7) ++
-  new constellation.rc.WithSbusNoCInNodeMapping ("Core 4"   , 11) ++
-  new constellation.rc.WithSbusNoCInNodeMapping ("Core 3"   , 10) ++
-  new constellation.rc.WithSbusNoCInNodeMapping ("Core 2"   , 9) ++
-  new constellation.rc.WithSbusNoCInNodeMapping ("Core 1"   , 8) ++
-  new constellation.rc.WithSbusNoCInNodeMapping ("Core 0"   , 7) ++
-  new constellation.rc.WithSbusNoCInNodeMapping ("serial-tl", 0) ++
-
-  // Replace the XBar-based sbus with a NoC based sbus
-  new constellation.rc.WithSbusNoC ++
-
-  new chipyard.config.WithSystemBusWidth(256) ++
-
-  // Cores 1-4 are standard processing cores
-  new freechips.rocketchip.subsystem.WithNBigCores(4) ++
-
-  // Core 0 is small control core, minimize cache
-  new freechips.rocketchip.subsystem.WithL1ICacheSets(64) ++
-  new freechips.rocketchip.subsystem.WithL1ICacheWays(1) ++
-  new freechips.rocketchip.subsystem.WithL1DCacheSets(64) ++
-  new freechips.rocketchip.subsystem.WithL1DCacheWays(1) ++
-  new freechips.rocketchip.subsystem.WithNBigCores(1) ++
-
-  // 4 L2 banks
-  new freechips.rocketchip.subsystem.WithNBanks(4)++
-
-
-  // 4 DRAM channels
+// DOC include start: MultiNoCConfig
+class MultiNoCConfig extends Config(
+  new constellation.soc.WithCbusNoC(constellation.protocol.TLNoCParams(
+    constellation.protocol.DiplomaticNetworkNodeMapping(
+      inNodeMapping = ListMap(
+        "serial-tl" -> 0),
+      outNodeMapping = ListMap(
+        "error" -> 1, "l2[0]" -> 2, "pbus" -> 3, "plic" -> 4,
+        "clint" -> 5, "dmInner" -> 6, "bootrom" -> 7, "tileClockGater" -> 8, "tileResetSetter" -> 9)),
+    NoCParams(
+      topology = TerminalRouter(BidirectionalLine(10)),
+      channelParamGen = (a, b) => UserChannelParams(Seq.fill(5) { UserVirtualChannelParams(4) }),
+      routingRelation = NonblockingVirtualSubnetworksRouting(TerminalRouterRouting(BidirectionalLineRouting()), 5, 1))
+  )) ++
+  new constellation.soc.WithMbusNoC(constellation.protocol.TLNoCParams(
+    constellation.protocol.DiplomaticNetworkNodeMapping(
+      inNodeMapping = ListMap(
+        "L2 InclusiveCache[0]" -> 1, "L2 InclusiveCache[1]" -> 2,
+        "L2 InclusiveCache[2]" -> 5, "L2 InclusiveCache[3]" -> 6),
+      outNodeMapping = ListMap(
+        "system[0]" -> 0, "system[1]" -> 3,  "system[2]" -> 4 , "system[3]" -> 7,
+        "serdesser" -> 0)),
+    NoCParams(
+      topology        = TerminalRouter(BidirectionalTorus1D(8)),
+      channelParamGen = (a, b) => UserChannelParams(Seq.fill(10) { UserVirtualChannelParams(4) }),
+      routingRelation = BlockingVirtualSubnetworksRouting(TerminalRouterRouting(BidirectionalTorus1DShortestRouting()), 5, 2))
+  )) ++
+  new constellation.soc.WithSbusNoC(constellation.protocol.TLNoCParams(
+    constellation.protocol.DiplomaticNetworkNodeMapping(
+      inNodeMapping = ListMap(
+        "Core 0" -> 1, "Core 1" -> 2,  "Core 2" -> 4 , "Core 3" -> 7,
+        "Core 4" -> 8, "Core 5" -> 11, "Core 6" -> 13, "Core 7" -> 14,
+        "serial-tl" -> 0),
+      outNodeMapping = ListMap(
+        "system[0]" -> 5, "system[1]" -> 6, "system[2]" -> 9, "system[3]" -> 10,
+        "pbus" -> 3)),
+    NoCParams(
+      topology        = TerminalRouter(Mesh2D(4, 4)),
+      channelParamGen = (a, b) => UserChannelParams(Seq.fill(8) { UserVirtualChannelParams(4) }),
+      routingRelation = BlockingVirtualSubnetworksRouting(TerminalRouterRouting(Mesh2DEscapeRouting()), 5, 1))
+  )) ++
+  new freechips.rocketchip.subsystem.WithNBigCores(8) ++
+  new freechips.rocketchip.subsystem.WithNBanks(4) ++
   new freechips.rocketchip.subsystem.WithNMemoryChannels(4) ++
+  new chipyard.config.AbstractConfig
+)
+// DOC include end: MultiNoCConfig
 
-  new chipyard.config.AbstractConfig)
+/*
+ * 10 - 11 - 12 - 13 - 14
+ *            |
+ *      0 --- 1 --- 2 --- 3
+ *      |                 |
+ *      9                 4
+ *      |                 |
+ *      8 --- 7 --- 6 --- 5
+ *            |
+ * 15 - 16 - 17 - 18 - 19
+ *
+ * SI/SO: Inward/outward names into sbus
+ * MI/MO: Inward/outward names into mbus
+ *
+ * Agent  | Bus | String     | node
+ * ================================
+ * Core 0 | SI  | Core 0     |    2
+ * Core 1 | SI  | Core 1     |   10
+ * Core 2 | SI  | Core 2     |   11
+ * Core 3 | SI  | Core 3     |   13
+ * Core 4 | SI  | Core 4     |   14
+ * Core 5 | SI  | Core 5     |   15
+ * Core 6 | SI  | Core 6     |   16
+ * Core 7 | SI  | Core 7     |   18
+ * Core 8 | SI  | Core 8     |   19
+ * fbus   | SI  | serial-tl  |    9
+ * pbus   | SO  | pbus       |    4
+ * L2 0   | SO  | system[0]  |    0
+ * L2 1   | SO  | system[1]  |    2
+ * L2 2   | SO  | system[2]  |    8
+ * L2 3   | SO  | system[3]  |    6
+ * L2 0   | MI  | Cache[0]   |    0
+ * L2 1   | MI  | Cache[1]   |    2
+ * L2 2   | MI  | Cache[2]   |    8
+ * L2 3   | MI  | Cache[3]   |    6
+ * DRAM 0 | MO  | system[0]  |    3
+ * DRAM 1 | MO  | system[1]  |    5
+ * extram | MO  | serdesser  |    9
+ */
+// DOC include start: SharedNoCConfig
+class SharedNoCConfig extends Config(
+  new constellation.soc.WithGlobalNoC(GlobalNoCParams(
+    NoCParams(
+      topology        = TerminalRouter(HierarchicalTopology(
+        base     = UnidirectionalTorus1D(10),
+        children = Seq(HierarchicalSubTopology(1, 2, BidirectionalLine(5)),
+                       HierarchicalSubTopology(7, 2, BidirectionalLine(5))))),
+      channelParamGen = (a, b) => UserChannelParams(Seq.fill(22) { UserVirtualChannelParams(4) }),
+      routingRelation = NonblockingVirtualSubnetworksRouting(TerminalRouterRouting(HierarchicalRouting(
+        baseRouting = UnidirectionalTorus1DDatelineRouting(),
+        childRouting = Seq(BidirectionalLineRouting(),
+                           BidirectionalLineRouting()))), 10, 2)
+    )
+  )) ++
+  new constellation.soc.WithMbusNoC(constellation.protocol.TLNoCParams(
+    constellation.protocol.DiplomaticNetworkNodeMapping(
+      inNodeMapping = ListMap(
+        "Cache[0]" -> 0, "Cache[1]" -> 2, "Cache[2]" -> 8, "Cache[3]" -> 6),
+      outNodeMapping = ListMap(
+        "system[0]" -> 3, "system[1]" -> 5,
+        "serdesser" -> 9))
+  ), true) ++
+  new constellation.soc.WithSbusNoC(constellation.protocol.TLNoCParams(
+    constellation.protocol.DiplomaticNetworkNodeMapping(
+      inNodeMapping = ListMap(
+        "serial-tl" -> 9, "Core 0" -> 2,
+        "Core 1" -> 10, "Core 2" -> 11, "Core 3" -> 13, "Core 4" -> 14,
+        "Core 5" -> 15, "Core 6" -> 16, "Core 7" -> 18, "Core 8" -> 19),
+      outNodeMapping = ListMap(
+        "system[0]" -> 0, "system[1]" -> 2, "system[2]" -> 8, "system[3]" -> 6,
+        "pbus" -> 4))
+  ), true) ++
+  new freechips.rocketchip.subsystem.WithNBigCores(8) ++
+  new freechips.rocketchip.subsystem.WithNBanks(4) ++
+  new freechips.rocketchip.subsystem.WithNMemoryChannels(2) ++
+  new chipyard.config.AbstractConfig
+)
+// DOC include end: SharedNoCConfig