+++ generators/chipyard/src/main/scala/example/dsptools/StreamingPassthrough.scala
@@ -16,8 +16,8 @@ import freechips.rocketchip.subsystem._
 // Simple passthrough to use as testbed sanity check
 // StreamingPassthrough params
 case class StreamingPassthroughParams(
-  writeAddress: BigInt = 0x2000,
-  readAddress: BigInt = 0x2100,
+  writeAddress: BigInt = 0x2200,
+  readAddress: BigInt = 0x2300,
   depth: Int
 )